Basic Accelerometer Code
The demo software for the accelerometer is totally broken.  This
program is a basic "faking it" sort of implementation that is based on
the idea:
How many units in the binary range are used for 1G change?
Based on starting point of midrange (512 out of 1024) and known device
range (+/- 16g) where does that put the min and max value?

The answer to those questions results in calculating a range.  Range
then mapped int0 -1600 to +1600 range.  Divide by 100 to translate to
the Gs.